requirements.txt
python-dotenv
numpy
pandas
pymongo
pymongo[srv]
certifi
scikit-learn
dill
pyaml
# -e .